% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanning_tree.R
\name{fundamental_cuts}
\alias{fundamental_cuts}
\title{Fundamental cuts of a spanning tree}
\usage{
fundamental_cuts(graph, tree)
}
\arguments{
\item{graph}{An \code{adj} object representing a connected undirected graph.}

\item{tree}{An \code{adj} object representing a spanning tree of \code{graph}.}
}
\value{
A list of \code{adj} objects, one per tree edge. Each \code{adj} represents
the cut as a subgraph of \code{graph}.
}
\description{
Returns the fundamental cuts of a spanning tree with respect to the original
graph. There is one fundamental cut per tree edge (n - 1 total), consisting
of all edges in the graph that cross the bipartition induced by removing that
tree edge.
}
\examples{
g <- adj::adj(list(c(2L, 3L), c(1L, 3L), c(1L, 2L)))
t <- minimum_spanning_tree(g)
fundamental_cuts(g, t)
}
\seealso{
\code{\link[=fundamental_cycles]{fundamental_cycles()}}
}
