% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanning_tree.R
\name{is_spanning_tree}
\alias{is_spanning_tree}
\title{Test if a graph is a spanning tree}
\usage{
is_spanning_tree(tree)
}
\arguments{
\item{tree}{An \code{adj} object to test.}
}
\value{
A logical scalar.
}
\description{
Returns \code{TRUE} if \code{tree} is a spanning tree: a connected acyclic graph with
exactly n-1 edges. To additionally check that \code{tree} uses only edges from a
reference graph, see \code{\link[=is_spanning_tree_of]{is_spanning_tree_of()}}.
}
\examples{
g <- adj::adj(list(c(2L, 3L), c(1L, 3L), c(1L, 2L)))
t <- minimum_spanning_tree(g)
is_spanning_tree(t)
}
\seealso{
\code{\link[=is_spanning_tree_of]{is_spanning_tree_of()}}, \code{\link[=minimum_spanning_tree]{minimum_spanning_tree()}},
\code{\link[=enumerate_spanning_trees]{enumerate_spanning_trees()}}
}
