% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanning_tree.R
\name{is_spanning_tree_of}
\alias{is_spanning_tree_of}
\title{Test if a graph is a spanning tree of another graph}
\usage{
is_spanning_tree_of(tree, graph)
}
\arguments{
\item{tree}{An \code{adj} object to test.}

\item{graph}{An \code{adj} object representing a connected undirected graph.}
}
\value{
A logical scalar.
}
\description{
Returns \code{TRUE} if \code{tree} is a spanning tree of \code{graph}: a connected acyclic
subgraph that spans all vertices and uses only edges present in \code{graph}.
}
\examples{
g <- adj::adj(list(c(2L, 3L), c(1L, 3L), c(1L, 2L)))
t <- minimum_spanning_tree(g)
is_spanning_tree_of(t, g)
}
\seealso{
\code{\link[=is_spanning_tree]{is_spanning_tree()}}, \code{\link[=minimum_spanning_tree]{minimum_spanning_tree()}},
\code{\link[=enumerate_spanning_trees]{enumerate_spanning_trees()}}
}
